add custom fonts, use them in TimeComponent

+++ Sources/CalendarClock/Core/Constants.swift
@@ -1,3 +1,3 @@
-let SCREEN_WIDTH: Int32 = 1024; // px
-let SCREEN_HEIGHT: Int32 = 768; // px
-let CONTENT_HEIGHT: Int32 = 128; // px
+let SCREEN_WIDTH: Float = 1024; // px
+let SCREEN_HEIGHT: Float = 768; // px
+let CONTENT_HEIGHT: Float = 128; // px
